+++ src/utils/common.h
@@ -39,6 +39,6 @@ typedef uint64_t uint64;
 #define DSIM_INVALID_INDEX ((uint32_t)-1)
 
 uint64_t dsim_hash64( uint64_t key );
-size_t dsim_next_pow_2( size_t value );
+uint64_t dsim_next_pow_2( uint64_t value );
 
 DSIM_END_HEADER
